Create scrape_wiktionary.php
This code can successfully pull off the first loop, but the following loops don't progress to the next page per $nextURL. 

For $i = 1 or 2, the variable $nextURL is indeed a string containing the proper next URL, but the $data coming out of file_get_contents() is the first page of uncountable nouns. Strangely, on loops $i > 0, $data always reverts to the first page of uncountable nouns, even if you begin with the 2nd or 3rd page (hard coded into line 3).