Update scrape_wiktionary.php
Here is a more brute-force approach, creating my own search function to get the appropriate link. We were get a portion of a link at the start of the web page, and the ending of the link we wanted, leaving a large portion of unwanted code in the middle. I told the program that if it found a url link that was more than 300 characters long, then it probably was not a link, and asked it to keep searching where it left off.

I also was having problems with $data changing all '&' symbols to '&amp;' which was causing additional glitches. I did a str_replace that would remove all "amp;" portions of the code. It seemed to work. More code is to come, by putting all the words into a file where we can  start using the words. Put this was a good stopping point.